Develop

See merge request projetm2/backend!21

+++ src/Union.Backend.Model/Models/Wallet.cs
@@ -4,7 +4,7 @@ namespace Union.Backend.Model.Models
 {
     public class Wallet : UniqueEntity
     {
-        public int Balance { get; set; }
+        public float Balance { get; set; }
         public Guid OfUser { get; set; }
     }
 }
